Split thread.tsx into components, rename Composer to PromptInput, and add chat runtime ACL
