+++ src/components/assistant-ui/thread/index.ts
@@ -1,0 +1,1 @@
+export { Thread } from "./Thread";
