+++ src/components/assistant-ui/thread/message-components.ts
@@ -1,0 +1,9 @@
+import { AssistantMessage } from "./AssistantMessage";
+import { EditPromptInput } from "./EditPromptInput";
+import { UserMessage } from "./UserMessage";
+
+export const MESSAGE_COMPONENTS = {
+  UserMessage,
+  EditComposer: EditPromptInput,
+  AssistantMessage,
+};
