+++ src/lib/chat/runtime/index.ts
@@ -1,0 +1,11 @@
+export type {
+  ChatMessage as ChatMessageData,
+  ChatMessagePart,
+  ChatMessageRole,
+  ChatTextPart,
+  ComposerActions,
+  ComposerStateSnapshot,
+  ThreadState,
+} from "./types";
+export * from "./primitives";
+export * from "./hooks";
