+++ src/lib/chat/runtime/primitives.ts
@@ -1,0 +1,20 @@
+import {
+  ActionBarPrimitive,
+  AuiIf,
+  BranchPickerPrimitive,
+  ComposerPrimitive,
+  ErrorPrimitive,
+  MessagePrimitive,
+  ThreadPrimitive,
+} from "@assistant-ui/react";
+import type { BranchPickerPrimitive as _BPP } from "@assistant-ui/react";
+
+export const ChatThread = ThreadPrimitive;
+export const ChatMessage = MessagePrimitive;
+export const ChatPromptInput = ComposerPrimitive;
+export const ChatActionBar = ActionBarPrimitive;
+export const ChatBranchPicker = BranchPickerPrimitive;
+export const ChatError = ErrorPrimitive;
+export const ChatIf = AuiIf;
+
+export type ChatBranchPickerRootProps = _BPP.Root.Props;
